Fixes To Transformation And GeoMath

+++ src/data_generic_state_item/state_global_velocity.h
@@ -8,10 +8,10 @@ class StateGlobalVelocity
 public:
     StateGlobalVelocity();
 public:
-    double x;
-    double y;
-    double z;
-    double heading;
+    float x;
+    float y;
+    float z;
+    float heading;
 };
 
 } //end of namespace DataState
